requirements/dev.txt
# base requirements
-r base.txt

# test requirements
coverage
django-debug-toolbar
flake8
mock
pyaml
